+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * [PR-15](https://github.com/ITK-Leantime/data-api/pull/15)
   * Added timesheetTotals endpoint returning logged hours grouped by day or week.
+  * Added workDate filter (year or year-month) to the timesheetTotals endpoint.
 
 ## [0.1.2] - 2026-03-06